requirements.txt
aiohttp==3.8.5
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
azure-ai-textanalytics==5.3.0
#python-dotenv==1.0.0
python-dotenv>=0.19.0